Changes to arxiv/auth/legacy/tests
Move setup of db out to a mixin under arxiv.auth.legacy.tests.util
