+++ arxiv/auth/legacy/tests/test_accounts.py
@@ -1,21 +1,13 @@
 """Tests for :mod:`arxiv.users.legacy.accounts`."""
 
-import tempfile
-from datetime import datetime
-import shutil
-import hashlib
-from pytz import UTC
 from unittest import TestCase
-from sqlalchemy import select
-from flask import Flask
 
-from arxiv.config import Settings
 from arxiv.taxonomy import definitions
 from arxiv.db import transaction
 from arxiv.db import models
+from .util import SetUpUserMixin
 
-from .util import temporary_db
-from .. import util, authenticate, exceptions
+from .. import authenticate, exceptions
 from .. import accounts
 from ... import domain
 
@@ -35,84 +27,6 @@ def get_user(session, user_id):
         .first()
 
     return db_user, db_nick, db_profile
-
-
-class SetUpUserMixin(TestCase):
-    """Mixin for creating a test user and other database goodies."""
-
-    def setUp(self):
-        """Set up the database."""
-        self.db_path = tempfile.mkdtemp()
-        self.db_uri = f'sqlite:///{self.db_path}/test.db'
-        self.user_id = '15830'
-        self.app = Flask('test')
-        self.app.config['CLASSIC_SESSION_HASH'] = 'foohash'
-        self.app.config['CLASSIC_COOKIE_NAME'] = 'tapir_session_cookie'
-        self.app.config['SESSION_DURATION'] = '36000'
-        settings = Settings(
-                        CLASSIC_DB_URI=self.db_uri,
-                        LATEXML_DB_URI=None)
-
-        self.engine, _ = models.configure_db(settings)            # Insert tapir policy classes
-
-        with temporary_db("sqlite:///:memory:", drop=False) as session:
-            self.user_class = session.scalar(
-                select(models.TapirPolicyClass).where(models.TapirPolicyClass.class_id==2))
-            self.email = 'first@last.iv'
-            self.db_user = models.TapirUser(
-                user_id=self.user_id,
-                first_name='first',
-                last_name='last',
-                suffix_name='iv',
-                email=self.email,
-                policy_class=self.user_class.class_id,
-                flag_edit_users=1,
-                flag_email_verified=1,
-                flag_edit_system=0,
-                flag_approved=1,
-                flag_deleted=0,
-                flag_banned=0,
-                tracking_cookie='foocookie',
-            )
-            self.username = 'foouser'
-            self.db_nick = models.TapirNickname(
-                nickname=self.username,
-                user_id=self.user_id,
-                user_seq=1,
-                flag_valid=1,
-                role=0,
-                policy=0,
-                flag_primary=1
-            )
-            self.salt = b'foo'
-            self.password = b'thepassword'
-            hashed = hashlib.sha1(self.salt + b'-' + self.password).digest()
-            self.db_password = models.TapirUsersPassword(
-                user_id=self.user_id,
-                password_storage=2,
-                password_enc=hashed
-            )
-            n = util.epoch(datetime.now(tz=UTC))
-            self.secret = 'foosecret'
-            self.db_token = models.TapirPermanentToken(
-                user_id=self.user_id,
-                secret=self.secret,
-                valid=1,
-                issued_when=n,
-                issued_to='127.0.0.1',
-                remote_host='foohost.foo.com',
-                session_id=0
-            )
-            session.add(self.user_class)
-            session.add(self.db_user)
-            session.add(self.db_password)
-            session.add(self.db_nick)
-            session.add(self.db_token)
-            session.commit()
-
-    def tearDown(self):
-        """Drop tables from the in-memory db file."""
-        util.drop_all(self.engine)
 
 
 class TestUsernameExists(SetUpUserMixin):
