+++ arxiv/auth/legacy/tests/test_authenticate.py
@@ -17,86 +17,86 @@ from arxiv.db import models
 
 from .. import authenticate, exceptions, util
 
-from .util import temporary_db
+from .util import SetUpUserMixin
 
 from ..passwords import hash_password
 
 EASTERN = timezone('US/Eastern')
 
 
-class TestAuthenticateWithPermanentToken(TestCase):
+class TestAuthenticateWithPermanentToken(SetUpUserMixin):
     """User has a permanent token."""
 
-    def setUp(self):
-        """Instantiate a user and its credentials in the DB."""
-        self.db = f'sqlite:///:memory:'
-        self.user_id = '1'
-
-        self.app = Flask('test')
-        settings = Settings(
-                        CLASSIC_DB_URI=self.db,
-                        LATEXML_DB_URI=None)
-
-        self.engine, _ = models.configure_db(settings)
-
-        with temporary_db(self.db, drop=False) as session:
-            self.user_class = session.scalar(
-                select(models.TapirPolicyClass).where(models.TapirPolicyClass.class_id==2))
-            self.email = 'first@last.iv'
-            self.db_user = models.TapirUser(
-                user_id=self.user_id,
-                first_name='first',
-                last_name='last',
-                suffix_name='iv',
-                email=self.email,
-                policy_class=self.user_class.class_id,
-                flag_edit_users=1,
-                flag_email_verified=1,
-                flag_edit_system=0,
-                flag_approved=1,
-                flag_deleted=0,
-                flag_banned=0,
-                tracking_cookie='foocookie',
-            )
-            self.username = 'foouser'
-            self.db_nick = models.TapirNickname(
-                nickname=self.username,
-                user_id=self.user_id,
-                user_seq=1,
-                flag_valid=1,
-                role=0,
-                policy=0,
-                flag_primary=1
-            )
-            self.salt = b'foo'
-            self.password = b'thepassword'
-            hashed = hashlib.sha1(self.salt + b'-' + self.password).digest()
-            self.db_password = models.TapirUsersPassword(
-                user_id=self.user_id,
-                password_storage=2,
-                password_enc=hashed
-            )
-            n = util.epoch(datetime.now(tz=UTC))
-            self.secret = 'foosecret'
-            self.db_token = models.TapirPermanentToken(
-                user_id=self.user_id,
-                secret=self.secret,
-                valid=1,
-                issued_when=n,
-                issued_to='127.0.0.1',
-                remote_host='foohost.foo.com',
-                session_id=0
-            )
-            session.add(self.user_class)
-            session.add(self.db_user)
-            session.add(self.db_password)
-            session.add(self.db_nick)
-            session.add(self.db_token)
-            session.commit()
-
-    def tearDown(self):
-        """Drop tables from the in-memory db file."""
-        util.drop_all(self.engine)
+    #def setUp(self):
+    #     """Instantiate a user and its credentials in the DB."""
+    #     self.db = f'sqlite:///:memory:'
+    #     self.user_id = '1'
+    #
+    #     self.app = Flask('test')
+    #     settings = Settings(
+    #                     CLASSIC_DB_URI=self.db,
+    #                     LATEXML_DB_URI=None)
+    #
+    #     self.engine, _ = models.configure_db(settings)
+    #
+    #     with temporary_db(self.db, drop=False) as session:
+    #         self.user_class = session.scalar(
+    #             select(models.TapirPolicyClass).where(models.TapirPolicyClass.class_id==2))
+    #         self.email = 'first@last.iv'
+    #         self.db_user = models.TapirUser(
+    #             user_id=self.user_id,
+    #             first_name='first',
+    #             last_name='last',
+    #             suffix_name='iv',
+    #             email=self.email,
+    #             policy_class=self.user_class.class_id,
+    #             flag_edit_users=1,
+    #             flag_email_verified=1,
+    #             flag_edit_system=0,
+    #             flag_approved=1,
+    #             flag_deleted=0,
+    #             flag_banned=0,
+    #             tracking_cookie='foocookie',
+    #         )
+    #         self.username = 'foouser'
+    #         self.db_nick = models.TapirNickname(
+    #             nickname=self.username,
+    #             user_id=self.user_id,
+    #             user_seq=1,
+    #             flag_valid=1,
+    #             role=0,
+    #             policy=0,
+    #             flag_primary=1
+    #         )
+    #         self.salt = b'foo'
+    #         self.password = b'thepassword'
+    #         hashed = hashlib.sha1(self.salt + b'-' + self.password).digest()
+    #         self.db_password = models.TapirUsersPassword(
+    #             user_id=self.user_id,
+    #             password_storage=2,
+    #             password_enc=hashed
+    #         )
+    #         n = util.epoch(datetime.now(tz=UTC))
+    #         self.secret = 'foosecret'
+    #         self.db_token = models.TapirPermanentToken(
+    #             user_id=self.user_id,
+    #             secret=self.secret,
+    #             valid=1,
+    #             issued_when=n,
+    #             issued_to='127.0.0.1',
+    #             remote_host='foohost.foo.com',
+    #             session_id=0
+    #         )
+    #         session.add(self.user_class)
+    #         session.add(self.db_user)
+    #         session.add(self.db_password)
+    #         session.add(self.db_nick)
+    #         session.add(self.db_token)
+    #         session.commit()
+    #
+    # def tearDown(self):
+    #     """Drop tables from the in-memory db file."""
+    #     util.drop_all(self.engine)
 
     def test_token_is_malformed(self):
         """Token is present, but it has the wrong format."""
@@ -141,69 +141,8 @@ class TestAuthenticateWithPermanentToken(TestCase):
                              "authorizations are set")
 
 
-class TestAuthenticateWithPassword(TestCase):
+class TestAuthenticateWithPassword(SetUpUserMixin):
     """User is attempting login with username+password."""
-
-    def setUp(self):
-        """Instantiate a user."""
-        self.db = f'sqlite:///:memory:'
-
-        self.app = Flask('test')
-        settings = Settings(
-                        CLASSIC_DB_URI=self.db,
-                        LATEXML_DB_URI=None)
-
-        self.engine, _ = models.configure_db(settings)
-        self.user_id = '5'
-        with temporary_db(self.db, drop=False) as session:
-            # We have a good old-fashioned user.
-            self.user_class = session.scalar(
-                select(models.TapirPolicyClass).where(models.TapirPolicyClass.class_id==2))
-            self.email = 'first@last.iv'
-            self.db_user = models.TapirUser(
-                user_id=self.user_id,
-                first_name='first',
-                last_name='last',
-                suffix_name='iv',
-                email=self.email,
-                policy_class=self.user_class.class_id,
-                flag_edit_users=1,
-                flag_email_verified=1,
-                flag_edit_system=0,
-                flag_approved=1,
-                flag_deleted=0,
-                flag_banned=0,
-                tracking_cookie='foocookie',
-            )
-            self.username = 'foouser'
-            self.db_nick = models.TapirNickname(
-                nickname=self.username,
-                user_id=self.user_id,
-                user_seq=1,
-                flag_valid=1,
-                role=0,
-                policy=0,
-                flag_primary=1
-            )
-            self.salt = b'foo'
-            self.password = 'thepassword'
-            # hashed = hashlib.sha1(self.salt + b'-' + self.password).digest()
-
-            hashed = hash_password(self.password)
-            self.db_password = models.TapirUsersPassword(
-                user_id=self.user_id,
-                password_storage=2,
-                password_enc=hashed
-            )
-            session.add(self.user_class)
-            session.add(self.db_user)
-            session.add(self.db_password)
-            session.add(self.db_nick)
-            session.commit()
-
-    def tearDown(self):
-        """Drop tables from the in-memory db file."""
-        util.drop_all(self.engine)
 
     def test_no_username(self):
         """Username is not entered."""
